requirements.txt
Pillow==11.3.0
PyQt5==5.15.11
PyQt5_sip==12.17.0
